Changes to allow generic snd_raw and send_address
