+++ library.properties
@@ -1,5 +1,5 @@
 name=IRremoteESP8266
-version=1.0.0
+version=2.0.0
 author=Sebastien Warin, Mark Szabo, Ken Shirriff
 maintainer=Sebastien Warin
 sentence=Send and receive infrared signals with multiple protocols.
